Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,7 +2,7 @@
 
 
 int main() {
-    int a,b,c;
+    float a,b,c;
 
     scanf("%d",&a);
     scanf("%d",&b);
